Use InMemSettingsProvider in other tests
