+++ collect_app/src/test/java/org/odk/collect/android/application/initialization/ExistingSettingsMigratorTest.kt
@@ -6,11 +6,9 @@ import org.hamcrest.Matchers.nullValue
 import org.junit.Test
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
-import org.odk.collect.android.preferences.source.SettingsProvider
+import org.odk.collect.android.support.InMemSettingsProvider
 import org.odk.collect.projects.InMemProjectsRepository
 import org.odk.collect.projects.Project
-import org.odk.collect.shared.Settings
-import org.odk.collect.testshared.InMemSettings
 
 class ExistingSettingsMigratorTest {
 
@@ -41,23 +39,4 @@ class ExistingSettingsMigratorTest {
         val existingSettingsMigrator = ExistingSettingsMigrator(mock(), mock(), mock())
         assertThat(existingSettingsMigrator.key(), `is`(nullValue()))
     }
-}
-
-private class InMemSettingsProvider : SettingsProvider {
-
-    private val metaSettings = InMemSettings()
-    private val settings = mutableMapOf<String?, InMemSettings>()
-
-    override fun getMetaSettings(): Settings {
-        return metaSettings
-    }
-
-    override fun getGeneralSettings(projectId: String?): Settings {
-        return settings.getOrPut(projectId) { InMemSettings() }
-    }
-
-    override fun getAdminSettings(projectId: String?): Settings {
-        return settings.getOrPut(projectId) { InMemSettings() }
-    }
-
 }
